// frontend/src/pages/careerForm/types.ts
export type FormData = {
  name: string
  age: string
  highschool_name: string
  highschool_stream: string
  college: string
  course_type: string
  course: string
  specialisation: string
  no_experience: boolean
  job_title: string
  company_name: string
  duration: string
  skills: string
  interests: string
  preferred_work_env: string
}